Rename `views` to `pages`
`pages` more accurately represents the purpose of the directory.
