+++ src/App.js
@@ -1,7 +1,7 @@
 import { MuiThemeProvider } from 'material-ui/styles'
 import { Provider as ReduxProvider } from 'react-redux'
 import { Route, BrowserRouter as Router } from 'react-router-dom'
-import EventsPage from 'views/EventsPage'
+import EventsPage from 'pages/EventsPage'
 import React from 'react'
 import Reboot from 'material-ui/Reboot'
 import injectSheet, { ThemeProvider } from 'react-jss'
