+++ src/pages/EventsPage.js
